fix: pst import (#34)
PSTs are stored as nodes and branches in NetworkData. When
removed during relevant branches scan, branches were upda-
ted, nodes were forgotten. This is fixed now.

Filtering by PST on relevant substations prevents important
use cases. We do not filter substations if they have a PST:
If we split one side of the grid and change PSTs in the other side.

* fix(importer): fix asset topology import where there are empty busbars with no elements
* refactor(importer): add relevant station rules to importer parameter
* test: test pst import
* test: add testcase to triggers pst import bug
* fix(dc_solver): PST branches and nodes are correctly filtered
* fix(dc_solver): Preprocess tests for pst node mask

Signed-off-by: Sascha Petznick <sascha.petznick@externel.be>